created parent class for settings
